updated the name with data-science-toolbox as per Jeroen
Hi Madhu,  Thanks! This looks great! Please be aware that my project is called "Data Science Toolbox" and not "Toolkit". The "Data Science Toolkit" is an older project by Pete Warden and is something different. If you're willing to rename your project then I would really appreciate it and it would also be less confusing for others.   Thanks!  Jeroen